Add an initial version of register safety checking.
This needs to be expanded to a general notion of "can this module be
flash-cleared?". This can be proved by checking the following:

1. ALl registers are reset only by some combination of module inputs
(usually just one reset signal). This is because reset signals that are
also controlled by registers mean that there's some delay when clearing
registers, which could be unsafe.
2. There exists a satisfiable assignment of the input signals that
simultaneously asserts the reset signals of all registers.