Need to compose transforms. SeqTransform handles it.
So the reset signal information is only available in HighForm/MidForm,
whereas the current dependency graph information only works in LowForm
(I believe this is because it doesn't want to handle a whole myriad of
types/partial connections).

So, we will have a pass that annotates DefRegisters
(firrtl.annotations.transformations.AddResetSignalInformation), which
will then feed into the original firrtl.analyses.CheckSpeculativeSafety
transform. The sequential combination is in
firrtl.analyses.VerifyModuleSquashability.