Pass now successfully uses the ResetSignalInfo.
Current conceptual issues:
1. Converting from middle to low form generates additional registers.
We need to figure out what these are.

Current implementation issues:
- Current implementation of the CheckSpeculativeSafety pass uses a TON
of memory. This is due to the lazy way that we're doing the analysis, I
believe. Maybe don't store multiple fields per register.

Current investigatory tasks:
- There are a lot of other passes that may do what we are doing better
(as in, there's a pass that removes the reset information, so we should
check and see how it works, etc).